chore: move logic to `vite:extendConfig` hook
