+++ src/utils/configure-vite.ts
@@ -1,5 +1,7 @@
 import type { Nuxt } from '@nuxt/schema'
 import defu from 'defu'
+import type { AssetURLOptions } from '@vue/compiler-sfc'
+import { transformAssetUrls as vuetifyTransformAssetUrls } from 'vite-plugin-vuetify'
 import { vuetifyStylesPlugin } from '../vite/vuetify-styles-plugin'
 import { vuetifyConfigurationPlugin } from '../vite/vuetify-configuration-plugin'
 import { vuetifyIconsPlugin } from '../vite/vuetify-icons-configuration-plugin'
@@ -8,8 +10,10 @@ import { vuetifySSRClientHintsPlugin } from '../vite/vuetify-ssr-client-hints-pl
 import { vuetifyImportPlugin } from '../vite/vuetify-import-plugin'
 import { checkVuetifyPlugins } from './module'
 import type { VuetifyNuxtContext } from './config'
+import { normalizeTransformAssetUrls } from './index'
 
 export function configureVite(configKey: string, nuxt: Nuxt, ctx: VuetifyNuxtContext) {
+  const { includeTransformAssetsUrls } = ctx.moduleOptions
   nuxt.hook('vite:extend', ({ config }) => checkVuetifyPlugins(config))
   nuxt.hook('vite:extendConfig', (viteInlineConfig) => {
     viteInlineConfig.plugins = viteInlineConfig.plugins || []
@@ -28,6 +32,39 @@ export function configureVite(configKey: string, nuxt: Nuxt, ctx: VuetifyNuxtCon
         ),
         configKey,
       ]
+    }
+
+    if (includeTransformAssetsUrls) {
+      nuxt.hook('vite:extendConfig', (config) => {
+        config.vue ??= {}
+        config.vue.template ??= {}
+        let existingTransformAssetUrls = config.vue?.template?.transformAssetUrls ?? {}
+        let useURLOptions: AssetURLOptions | undefined
+        if (typeof existingTransformAssetUrls === 'boolean') {
+          existingTransformAssetUrls = {}
+        }
+        else if ('base' in existingTransformAssetUrls || 'includeAbsolute' in existingTransformAssetUrls || 'tags' in existingTransformAssetUrls) {
+          useURLOptions = {
+            base: existingTransformAssetUrls.base as string | undefined,
+            includeAbsolute: existingTransformAssetUrls.includeAbsolute as boolean | undefined,
+          }
+          existingTransformAssetUrls = (existingTransformAssetUrls.tags ?? {}) as Record<string, string[]>
+        }
+
+        const transformAssetUrls = normalizeTransformAssetUrls(
+          typeof includeTransformAssetsUrls === 'object'
+            ? defu(existingTransformAssetUrls, vuetifyTransformAssetUrls, includeTransformAssetsUrls)
+            : defu(existingTransformAssetUrls, vuetifyTransformAssetUrls),
+        )
+
+        if (useURLOptions) {
+          useURLOptions.tags = transformAssetUrls
+          config.vue.template.transformAssetUrls = useURLOptions
+        }
+        else {
+          config.vue.template.transformAssetUrls = transformAssetUrls
+        }
+      })
     }
 
     viteInlineConfig.plugins.push(vuetifyImportPlugin({}))
